+++ src/main/java/techcourse/myblog/dto/user/UpdateUserRequest.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.dto;
+package techcourse.myblog.dto.user;
 
 import javax.persistence.Column;
 import javax.validation.constraints.Pattern;
