+++ src/main/java/techcourse/myblog/exception/comment/CommentAuthenticationException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.exception;
+package techcourse.myblog.exception.comment;
 
 public class CommentAuthenticationException extends RuntimeException {
