+++ src/main/java/techcourse/myblog/repository/ArticleRepository.java
@@ -3,7 +3,7 @@ package techcourse.myblog.repository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-import techcourse.myblog.domain.Article;
+import techcourse.myblog.domain.article.Article;
 
 public interface ArticleRepository extends JpaRepository<Article, Long> {
     Page<Article> findAll(Pageable pageable);
