+++ src/main/java/techcourse/myblog/repository/CommentRepository.java
@@ -1,8 +1,8 @@
 package techcourse.myblog.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import techcourse.myblog.domain.Article;
-import techcourse.myblog.domain.Comment;
+import techcourse.myblog.domain.article.Article;
+import techcourse.myblog.domain.comment.Comment;
 
 import java.util.List;
