+++ src/main/java/techcourse/myblog/repository/UserRepository.java
@@ -1,7 +1,7 @@
 package techcourse.myblog.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import techcourse.myblog.domain.User;
+import techcourse.myblog.domain.user.User;
 
 import java.util.Optional;
