+++ src/main/java/techcourse/myblog/service/ArticleService.java
@@ -5,13 +5,17 @@ import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import techcourse.myblog.domain.*;
-import techcourse.myblog.dto.ArticleRequest;
-import techcourse.myblog.dto.ArticleResponse;
-import techcourse.myblog.dto.CommentResponse;
-import techcourse.myblog.dto.UserResponse;
-import techcourse.myblog.exception.ArticleAuthenticationException;
-import techcourse.myblog.exception.ArticleException;
+import techcourse.myblog.domain.article.Article;
+import techcourse.myblog.domain.article.ArticleAssembler;
+import techcourse.myblog.domain.comment.Comment;
+import techcourse.myblog.domain.comment.CommentAssembler;
+import techcourse.myblog.domain.user.User;
+import techcourse.myblog.dto.article.ArticleRequest;
+import techcourse.myblog.dto.article.ArticleResponse;
+import techcourse.myblog.dto.comment.CommentResponse;
+import techcourse.myblog.dto.user.UserResponse;
+import techcourse.myblog.exception.article.ArticleAuthenticationException;
+import techcourse.myblog.exception.article.ArticleException;
 import techcourse.myblog.repository.ArticleRepository;
 import techcourse.myblog.utils.page.PageRequest;
