+++ src/main/java/techcourse/myblog/web/ArticleController.java
@@ -3,9 +3,9 @@ package techcourse.myblog.web;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import techcourse.myblog.dto.ArticleRequest;
-import techcourse.myblog.dto.ArticleResponse;
-import techcourse.myblog.dto.UserResponse;
+import techcourse.myblog.dto.article.ArticleRequest;
+import techcourse.myblog.dto.article.ArticleResponse;
+import techcourse.myblog.dto.user.UserResponse;
 import techcourse.myblog.service.ArticleService;
 import techcourse.myblog.utils.model.ModelUtil;
 import techcourse.myblog.utils.session.SessionUtil;
