+++ src/main/java/techcourse/myblog/web/UserController.java
@@ -9,10 +9,10 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
-import techcourse.myblog.dto.LoginRequest;
-import techcourse.myblog.dto.SignUpRequest;
-import techcourse.myblog.dto.UpdateUserRequest;
-import techcourse.myblog.dto.UserResponse;
+import techcourse.myblog.dto.user.LoginRequest;
+import techcourse.myblog.dto.user.SignUpRequest;
+import techcourse.myblog.dto.user.UpdateUserRequest;
+import techcourse.myblog.dto.user.UserResponse;
 import techcourse.myblog.service.LoginService;
 import techcourse.myblog.service.UserService;
 import techcourse.myblog.utils.model.ModelUtil;
