+++ src/test/java/techcourse/myblog/domain/ArticleTest.java
@@ -1,6 +1,8 @@
 package techcourse.myblog.domain;
 
 import org.junit.jupiter.api.Test;
+import techcourse.myblog.domain.article.Article;
+import techcourse.myblog.domain.user.User;
 
 import static org.assertj.core.api.Assertions.assertThat;
