+++ src/test/java/techcourse/myblog/repository/UserRepositoryTest.java
@@ -5,8 +5,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import techcourse.myblog.domain.User;
-import techcourse.myblog.exception.SignUpException;
+import techcourse.myblog.domain.user.User;
+import techcourse.myblog.exception.user.SignUpException;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
