+++ src/test/java/techcourse/myblog/service/UserServiceTest.java
@@ -5,13 +5,13 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import techcourse.myblog.domain.User;
-import techcourse.myblog.domain.UserAssembler;
-import techcourse.myblog.dto.SignUpRequest;
-import techcourse.myblog.dto.UpdateUserRequest;
-import techcourse.myblog.dto.UserResponse;
-import techcourse.myblog.exception.SignUpException;
-import techcourse.myblog.exception.UserException;
+import techcourse.myblog.domain.user.User;
+import techcourse.myblog.domain.user.UserAssembler;
+import techcourse.myblog.dto.user.SignUpRequest;
+import techcourse.myblog.dto.user.UpdateUserRequest;
+import techcourse.myblog.dto.user.UserResponse;
+import techcourse.myblog.exception.user.SignUpException;
+import techcourse.myblog.exception.user.UserException;
 import techcourse.myblog.repository.UserRepository;
 
 import java.util.Optional;
